Update everything to php 7.4
Updates the dependencies to 7.4 or greater, including the tests an phpunit.
Adds new typehints and updates the deprecated functions etc in the tests.

The current codeclimate-phan doesn't support type-hinting (throws syntax error)
which is probably caused by running phan on an older version of php. For now
we will skip phan in the codeclimate checks as the inner working of the codeclimate
docker are a bit murky for me.

This commit should have no functional changes to the working and therefore be
backwards-compatible with the previous versions.