Fixed some backward incompatible bugs/inconsistencies
Major possibly breaking change: the constructor has been made final.
This is because we directly call 'static()', which means we can not in
good faith depend on an override of the constructor.

Apart from that it adds some typehints to the decode function, that is
the same as the hints in json_decode, so should not give any issues. There
are also some parameters specified to help phpstan along.